Fixed configure.ac (again...) configure.ac had the server_interp_order and num_ghost_cells options deleted during the Nov 15, 2025 commit. **PLEASE pay more attention when removing blocks of code and test all changes thoroughly in an incremental fashion.** This was causing (silent) errors when attempting to use the server background types. These options have been added to the configuration script. In addition, the new compilation options threw a lot more warnings than before (eg. int vs long unsigned int comparison). Some files were edited to remove these warnings, although some warnings persist in parts of the code that are still in development (MHD stuff). There were a few orphaned c_code references in server_cartesian.cc that we missed after the original push to the new units convention. These now belong to the Fluid units. All benchmark tests run successfully. The parker spiral magnetic field test was enhanced to also compare with the discretized background version. The run_tests.sh script was improved by adding timing to the run portions of the tests. Also, the tests have been reorganized in order of CPUs required to run. As a point of comparison, script ran for all tests and 8 Intel i7-12700KF CPUs (including configuration, compilation, and plotting) took just under 9 minutes to complete. The longest runtime for a single test was 109 seconds for the PITCH ANGLE DISTRIBUTION ISOTROPIZATION benchmark.

+++ src/initial_base.hh
@@ -128,7 +128,7 @@ protected:
    std::vector <tableClass> initquant;
 
 //! Table entry counter
-   int table_counter;
+   long unsigned int table_counter;
 
 //! Set up the initial condition generator based on "params"
    void SetupInitial(bool construct) override;
